[LibEMU] -Fix: MSVC warnings about silent (u)int16 -> (u)int8 casting (MSVC doesn't like that :p)
git-svn-id: svn://svn.opendune.org/extra/libemu@196 43844f39-a229-4fad-ac6f-18a7c4f0fe92

+++ src/libemu.c
@@ -133,7 +133,7 @@ void emu_init(int argc, char **argv)
 		emu_last_ip   = values[16];
 		emu_last_crc  = values[17];
 		emu_last_length = values[18];
-		emu_overlay   = values[19];
+		emu_overlay   = (uint8)values[19];
 	}
 
 	/* Load the memory */
